# optimize to_byte_idx